dom/webgpu/tests/mochitest/test_gpu_disabled.html

<!doctype html>
<html>
  <head>
    <meta charset="utf-8" />
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" href="/tests/SimpleTest/test.css" />
  </head>
  <body>
    <script>
      ok(
        !SpecialPowers.getBoolPref("dom.webgpu.enabled"),
        "Pref should be disabled."
      );
      ok(navigator.gpu === undefined, "navigator.gpu === undefined");
      const canvas = document.createElement("canvas");
      ok(
        canvas.getContext("webgpu") === null,
        'canvas.getContext("webgpu") === null'
      );
    </script>
  </body>
</html>
